Refactor Ipopt tests
