+++ test/MOI_wrapper.jl
@@ -1,110 +1,144 @@
+module TestMOIWrapper
+
+using Ipopt
 using MathOptInterface
+using Test
+
 const MOI = MathOptInterface
-const MOIT = MOI.Test
-const MOIU = MOI.Utilities
-const MOIB = MOI.Bridges
+
+const OPTIMIZER = Ipopt.Optimizer()
+MOI.set(OPTIMIZER, MOI.Silent(), true)
 
 # Without fixed_variable_treatment set, duals are not computed for variables
 # that have lower_bound == upper_bound.
-const optimizer = Ipopt.Optimizer()
+MOI.set(
+    OPTIMIZER, MOI.RawParameter("fixed_variable_treatment"), "make_constraint"
+)
 
-MOI.set(optimizer, MOI.RawParameter("print_level"), 0)
-MOI.set(optimizer, MOI.RawParameter("fixed_variable_treatment"),
-        "make_constraint")
+# TODO(odow): add features to Ipopt so we can remove some of this caching.
+const BRIDGED_OPTIMIZER = MOI.Bridges.full_bridge_optimizer(
+    MOI.Utilities.CachingOptimizer(
+        MOI.Utilities.UniversalFallback(MOI.Utilities.Model{Float64}()),
+        OPTIMIZER,
+    ),
+    Float64,
+)
 
-const config = MOIT.TestConfig(atol=1e-4, rtol=1e-4,
-                               optimal_status=MOI.LOCALLY_SOLVED)
-# DualObjectiveValue is not implemented, so Ipopt does not pass the tests that
-# query it.
-# TODO: Consider implementing DualObjectiveValue for purely linear problems.
-const config_no_duals = MOIT.TestConfig(atol=1e-4, rtol=1e-4, duals=false,
-                                        optimal_status=MOI.LOCALLY_SOLVED)
+const CONFIG = MOI.Test.TestConfig(
+    atol = 1e-4,
+    rtol = 1e-4,
+    optimal_status = MOI.LOCALLY_SOLVED,
+    infeas_certificates = false,
+)
 
-@testset "SolverName" begin
-    @test MOI.get(optimizer, MOI.SolverName()) == "Ipopt"
+const CONFIG_NO_DUAL = MOI.Test.TestConfig(
+    atol = 1e-4,
+    rtol = 1e-4,
+    optimal_status = MOI.LOCALLY_SOLVED,
+    infeas_certificates = false,
+    duals = false,  # Don't check dual result!
+)
+
+function test_solvername()
+    @test MOI.get(OPTIMIZER, MOI.SolverName()) == "Ipopt"
 end
 
-@testset "supports_default_copy_to" begin
-    @test MOIU.supports_default_copy_to(optimizer, false)
-    @test !MOIU.supports_default_copy_to(optimizer, true)
+function test_supports_default_copy_to()
+    @test MOI.Utilities.supports_default_copy_to(OPTIMIZER, false)
+    @test !MOI.Utilities.supports_default_copy_to(OPTIMIZER, true)
 end
 
-@testset "Unit" begin
-    bridged = MOIB.full_bridge_optimizer(
-        Ipopt.Optimizer(print_level=0, fixed_variable_treatment="make_constraint"),
-        Float64)
-    # A number of test cases are excluded because loadfromstring! works only
-    # if the solver supports variable and constraint names.
-    exclude = ["delete_variable", # Deleting not supported.
-               "delete_variables", # Deleting not supported.
-               "getvariable", # Variable names not supported.
-               "solve_zero_one_with_bounds_1", # Variable names not supported.
-               "solve_zero_one_with_bounds_2", # Variable names not supported.
-               "solve_zero_one_with_bounds_3", # Variable names not supported.
-               "getconstraint", # Constraint names not suported.
-               "variablenames", # Variable names not supported.
-               "solve_with_upperbound", # loadfromstring!
-               "solve_with_lowerbound", # loadfromstring!
-               "solve_integer_edge_cases", # loadfromstring!
-               "solve_affine_lessthan", # loadfromstring!
-               "solve_affine_greaterthan", # loadfromstring!
-               "solve_affine_equalto", # loadfromstring!
-               "solve_affine_interval", # loadfromstring!
-               "get_objective_function", # Function getters not supported.
-               "solve_constant_obj",  # loadfromstring!
-               "solve_blank_obj", # loadfromstring!
-               "solve_singlevariable_obj", # loadfromstring!
-               "solve_objbound_edge_cases", # ObjectiveBound not supported.
-               "solve_affine_deletion_edge_cases", # Deleting not supported.
-               "solve_unbounded_model", # `NORM_LIMIT`
-               "number_threads", # NumberOfThreads not supported
-               "delete_nonnegative_variables", # get ConstraintFunction n/a.
-               "update_dimension_nonnegative_variables", # get ConstraintFunction n/a.
-               "delete_soc_variables", # VectorOfVar. in SOC not supported
-               "solve_result_index", # DualObjectiveValue not supported
-               ]
-    MOIT.unittest(bridged, config, exclude)
+function test_basicconstraint()
+    # TODO(odow): these basic constraint tests are broken.
+    # MOI.Test.basic_constraint_tests(OPTIMIZER, CONFIG)
 end
 
-@testset "MOI Linear tests" begin
-    exclude = ["linear8a", # Behavior in infeasible case doesn't match test.
-               "linear12", # Same as above.
-               "linear8b", # Behavior in unbounded case doesn't match test.
-               "linear8c", # Same as above.
-               "linear7",  # VectorAffineFunction not supported.
-               "linear15", # VectorAffineFunction not supported.
-               ]
-    model_for_ipopt = MOIU.UniversalFallback(MOIU.Model{Float64}())
-    linear_optimizer = MOI.Bridges.Constraint.SplitInterval{Float64}(
-                         MOIU.CachingOptimizer(model_for_ipopt, optimizer))
-    MOIT.contlineartest(linear_optimizer, config_no_duals, exclude)
-    # Tests setting bounds of `SingleVariable` constraint
-    MOIT.linear4test(optimizer, config_no_duals)
+function test_unittest()
+    MOI.Test.unittest(
+        BRIDGED_OPTIMIZER,
+        CONFIG,
+        String[
+            # VectorOfVariables-in-SecondOrderCone not supported
+            "delete_soc_variables",
+            # NumberOfThreads not supported
+            "number_threads",
+            # MOI.Integer not supported.
+            "solve_integer_edge_cases",
+            # ObjectiveBound not supported.
+            "solve_objbound_edge_cases",
+            # DualObjectiveValue not supported.
+            "solve_result_index",
+            # Returns NORM_LIMIT instead of DUAL_INFEASIBLE
+            "solve_unbounded_model",
+            # MOI.ZeroOne not supported.
+            "solve_zero_one_with_bounds_1",
+            "solve_zero_one_with_bounds_2",
+            "solve_zero_one_with_bounds_3",
+        ]
+    )
 end
 
-MOI.empty!(optimizer)
-
-@testset "MOI QP/QCQP tests" begin
-    qp_optimizer = MOIU.CachingOptimizer(MOIU.Model{Float64}(), optimizer)
-    MOIT.qptest(qp_optimizer, config)
-    exclude = ["qcp1", # VectorAffineFunction not supported.
-              ]
-    MOIT.qcptest(qp_optimizer, config_no_duals, exclude)
+function test_contlinear()
+    MOI.Test.contlineartest(
+        BRIDGED_OPTIMIZER,
+        CONFIG,
+        String[
+            # Tests requiring DualObjectiveValue. Tested below.
+            "linear1",
+            "linear2",
+            "linear10",
+            "linear14",
+            # Tests requiring infeasibility certificates
+            "linear8a",
+            "linear8b",
+            "linear8c",
+            "linear12",
+            # An INVALID_MODEL because it contains an empty 0 == 0 row.
+            "linear15",
+        ]
+    )
+    MOI.Test.linear1test(BRIDGED_OPTIMIZER, CONFIG_NO_DUAL)
+    MOI.Test.linear2test(BRIDGED_OPTIMIZER, CONFIG_NO_DUAL)
+    MOI.Test.linear10test(BRIDGED_OPTIMIZER, CONFIG_NO_DUAL)
+    MOI.Test.linear14test(BRIDGED_OPTIMIZER, CONFIG_NO_DUAL)
 end
 
-MOI.empty!(optimizer)
-
-@testset "MOI NLP tests" begin
-    MOIT.nlptest(optimizer, config)
+function test_qp()
+    MOI.Test.qptest(BRIDGED_OPTIMIZER, CONFIG)
 end
 
-@testset "Testing getters" begin
-    MOI.Test.copytest(MOI.instantiate(Ipopt.Optimizer, with_bridge_type=Float64), MOIU.Model{Float64}())
+function test_qcp()
+    MOI.empty!(BRIDGED_OPTIMIZER)
+    MOI.Test.qcptest(BRIDGED_OPTIMIZER, CONFIG_NO_DUAL)
 end
 
-@testset "Bounds set twice" begin
-    MOI.Test.set_lower_bound_twice(optimizer, Float64)
-    MOI.Test.set_upper_bound_twice(optimizer, Float64)
+function test_nlptest()
+    MOI.Test.nlptest(OPTIMIZER, CONFIG)
 end
 
-# TODO run basic tests for `SingleVariable` constraints: https://github.com/jump-dev/MathOptInterface.jl/pull/1133
+function test_getters()
+    MOI.Test.copytest(
+        MOI.instantiate(Ipopt.Optimizer, with_bridge_type=Float64),
+        MOI.Utilities.Model{Float64}()
+    )
+end
+
+function test_boundsettwice()
+    MOI.Test.set_lower_bound_twice(OPTIMIZER, Float64)
+    MOI.Test.set_upper_bound_twice(OPTIMIZER, Float64)
+end
+
+
+function test_nametest()
+    MOI.Test.nametest(BRIDGED_OPTIMIZER)
+end
+function test_validtest()
+    MOI.Test.validtest(BRIDGED_OPTIMIZER)
+end
+function test_emptytest()
+    MOI.Test.emptytest(BRIDGED_OPTIMIZER)
+end
+
+end  # module TestMOIWrapper
+
+runtests(TestMOIWrapper)
